+++ app/src/main/java/com/example/starfinder/models/StarInfo.kt
@@ -2,8 +2,8 @@ package com.example.starfinder.models
 
 data class StarInfo(
     val name: String,
-    val ra: Float,    // Прямое восхождение (в градусах)
-    val dec: Float,   // Склонение (в градусах)
+    val ra: Double,    // Прямое восхождение (в градусах)
+    val dec: Double,   // Склонение (в градусах)
     val epoch: String = "J2000",
     val dataSourceId: Int
 )
